requirements.txt
Flask
transformers
torch
requests
python-dotenv
gunicorn
numpy